+++ src/add-ons/kernel/busses/usb/xhci/xhci.h
@@ -477,7 +477,7 @@ private:
 			PciDevice*			fPciDevice {};
 
 			UsbStack*			fStack {};
-			uint8				fIRQ {};
+			uint32				fIRQ {};
 			bool				fUseMSI {};
 
 			area_id				fErstArea = -1;
